backend/requirements.txt
asyncio
asgi-redis
beautifulsoup4
channels
coreapi
coreschema
cryptography
Django==3.2.13
django-admin-tools
django-actions
django-admin-import
django-admin-sortable
django-allauth
django-appconf
django-channels
django-crontab
django-classy-tags
django-extensions
django-grappelli==3.0.3
django-online-users
django-imagekit 
django-user-accounts
django-clear-cache
django-cors-headers
django-redis
django-redis-sessions
django-guardian
django-reversion-rest-framework
django-registration-rest-framework
django-rest-auth
django-searchable-encrypted-fields
djangorestframework-bulk
drf-extensions
django-filter
django-admin-autocomplete-filter
django-rest-authtoken
djangorestframework-filters==1.0.0.dev2
djangorestframework==3.13.1
django-rest-swagger
djangorestframework-oauth
djangorestframework-jwt
django-filter==2.4.0
django-utils-six
flake8
msgpack-python
notifications
Pillow
PyYAML
Pygments
pyOpenSSL
pyparsing
pyres
pyflakes
python-dateutil
python-mimeparse
psycopg2
python-social-auth
python-redis-cache
redis
requests
requests-oauthlib
uWSGI
urllib3
six
